requirements.txt
arabic-reshaper==2.1.3
arrow==1.1.1
asgiref==3.4.1
autopep8==1.5.7
blessed==1.18.1
boto3==1.18.55
botocore==1.21.55
Brotli==1.0.9
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
croniter==1.0.15
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-advanced-password-validation==1.0.4
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-environ==0.4.5
django-filter==2.4.0
django-floppyforms==1.9.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-q==1.3.9
django-random-id-model==0.1.0
django-session-security==2.6.6
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
dnspython==1.16.0
et-xmlfile==1.1.0
fanstatic==1.1
fonttools==4.27.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
js.jquery==3.3.1
numpy==1.19.5
openpyxl==3.0.9
pandas==1.1.5
Pillow==8.4.0
pip-autoremove==0.9.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.1.1
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
reportlab==3.6.2
requests==2.25.1
s3transfer==0.5.0
shutilwhich==1.1.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==15.1.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
WebOb==1.8.7
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
zopfli==0.1.9
